requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
httpx
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
asyncpg>=0.27.0
alembic>=1.10.0
python-dotenv>=1.0.0
httpx>=0.24.0
psycopg2-binary>=2.9.6
cryptography>=41.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
pydantic-settings>=2.10.1
bcrypt==3.2.0
loguru==0.7.3
asgi-correlation-id>=4.2.0
streamlit==1.45.1
google-auth
google-api-python-client
openai
supabase
pydantic[email]
aiofiles
google-genai>=1.0.0
langchain-google-genai>=2.0.0
elasticsearch>=8.0.0
boto3
celery>=5.3.0
redis>=4.5.0
flower>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
google-auth
google-auth-oauthlib
fastapi-mail
Optional